git-svn-id: https://svn.code.sf.net/p/curses-sharp/code/trunk@4 52a65036-936b-447d-8c72-270fba66c20f

+++ CursesWrapper/attr.c
@@ -71,7 +71,7 @@
 */
 
 int
-wrap_attrset(unsigned long attrs)
+wrap_attrset(unsigned int attrs)
 {
 	return attrset(attrs);
 }
